Proposed clarification to Incentive Payouts
Point 3. of the description states:

"Grant the Exchequer (based on the analysis of user / usage data gathered during the program) the authority to cancel, or extend the incentive program for additional 3 month intervals, until such time as the Exchequer determines such incentives are no longer meaningful." 

To my reading, this opens up the possibility that a user would stake for a period (eg. 3 years) under the expectation that the rewards would continue for the entire length of their stake. However, with the current wording, it would appear that if the exchequer did not renew the program, that the rewards would cease. 

I therefore have proposed changed language which would prevent this from happening. I have proposed the addition:
"The incentives offered upon staking will survive changes introduced by the Exchequer at a later date. A staked position that would, for example, earn 30% based on the incentive structure that existed upon creation of the stake lock-up, should continue to earn that 30% until the stake lock-up expires."

Also, minor spelling mistake corrected.